# ai_test/backend/tests/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Web测试平台接口文档",
    "description": "FastAPI",
    "version": "0.0.1"
  },
  "paths": {
...
    "/api/pro/envs/{id}": {
      "get": {
        "tags": [
          "测试环境管理"
        ],
        "summary": "获取单个测试环境详情",
        "operationId": "get_env_api_pro_envs__id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestEnvSchemas"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "测试环境管理"
        ],
        "summary": "删除测试环境",
        "operationId": "delete_env_api_pro_envs__id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "测试环境管理"
        ],
        "summary": "修改测试环境",
        "operationId": "update_env_api_pro_envs__id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEnvForm"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestEnvSchemas"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TestEnvSchemas": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "环境id"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "环境名称"
          },
          "create_time": {
            "type": "string",
            "format": "date-time",
            "title": "Create Time",
            "description": "创建时间"
          },
          "host": {
            "type": "string",
            "title": "Host",
            "description": "环境地址"
          },
          "global_vars": {
            "type": "object",
            "title": "Global Vars",
            "description": "全局变量"
          },
          "project_id": {
            "type": "integer",
            "title": "Project Id",
            "description": "所属项目"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "create_time",
          "host",
          "global_vars",
          "project_id"
        ],
        "title": "TestEnvSchemas"
      },
...
      "UpdateEnvForm": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "环境名称"
          },
          "host": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Host",
            "description": "环境地址"
          },
          "global_vars": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Global Vars",
            "description": "全局变量"
          }
        },
        "type": "object",
        "title": "UpdateEnvForm",
        "description": "修改测试环境"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/users/token"
          }
        }
      }
    }
  }
}